# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: agentInterface API
  version: 1.2.2
  description: |
    Unified ventes API. Use /api/ventes with a type discriminator (AU_VENDEUR/DU_VENDEUR).
    Authentication uses email as username.
    Webhooks are emitted on stock changes and ventes creation.
servers:
  - url: http://localhost:8080
paths:
...
  /api/webhooks/subscriptions:
    post:
      tags: [Webhooks]
      summary: Créer une souscription webhook
      operationId: createWebhookSub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
    get:
      tags: [Webhooks]
      summary: Lister les souscriptions webhook actives
      operationId: listWebhookSubs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionResponse'
  /api/webhooks/subscriptions/{id}:
    delete:
      tags: [Webhooks]
      summary: Supprimer une souscription webhook
      operationId: deleteWebhookSub
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Supprimé
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateSubscriptionRequest:
      type: object
      properties:
        targetUrl:
          type: string
          format: uri
        eventTypes:
          type: array
          items:
            type: string
            enum: [STOCK_CHANGED, VENTE_AU_VENDEUR_CREATED, VENTE_DU_VENDEUR_CREATED]
      required: [targetUrl]
    SubscriptionResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        targetUrl: { type: string, format: uri }
        active: { type: boolean }
        eventTypes:
          type: array
          items:
            type: string
            enum: [STOCK_CHANGED, VENTE_AU_VENDEUR_CREATED, VENTE_DU_VENDEUR_CREATED]